Fix edge case
In cases where the previous decompressed data was longer than the
current, there would be no null terminator at the end of the newly
decompressed data. This would cause a portino of the previous data to be
printed erroniously. This commit stores an ending \0 immediately after
the decompressed data in the output buffer.